requirements.txt
chunknorris>=0.1.0
openai>=1.0.0
tiktoken>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
gradio>=4.0.0
